+++ GRAPH/EventDispatcher.h
@@ -60,12 +60,12 @@ namespace GRAPH
 
             inline std::vector<EventListener*>* getFixedPriorityListeners() const { return _fixedListeners; }
             inline std::vector<EventListener*>* getSceneGraphPriorityListeners() const { return _sceneGraphListeners; }
-            inline ssize_t getGt0Index() const { return _gt0Index; }
-            inline void setGt0Index(ssize_t index) { _gt0Index = index; }
+            inline int64 getGt0Index() const { return _gt0Index; }
+            inline void setGt0Index(int64 index) { _gt0Index = index; }
         private:
             std::vector<EventListener*>* _fixedListeners;
             std::vector<EventListener*>* _sceneGraphListeners;
-            ssize_t _gt0Index;
+            int64 _gt0Index;
         };
 
         void addEventListener(EventListener* listener);
