+++ GRAPH/Node.h
@@ -129,7 +129,7 @@ namespace GRAPH
         virtual void enumerateChildren(const std::string &name, std::function<bool(Node* node)> callback) const;
         virtual HObjectVector<Node*>& getChildren() { return _children; }
         virtual const HObjectVector<Node*>& getChildren() const { return _children; }
-        virtual ssize_t getChildrenCount() const;
+        virtual int64 getChildrenCount() const;
         virtual void setParent(Node* parent);
         virtual Node* getParent() { return _parent; }
         virtual const Node* getParent() const { return _parent; }
@@ -196,7 +196,7 @@ namespace GRAPH
         void stopActionByTag(int tag);
         void stopAllActionsByTag(int tag);
         Action* getActionByTag(int tag);
-        ssize_t getNumberOfRunningActions() const;
+        int64 getNumberOfRunningActions() const;
 
         virtual void setScheduler(Scheduler* scheduler);
         virtual Scheduler* getScheduler() { return _scheduler; }
@@ -287,7 +287,7 @@ namespace GRAPH
     protected:
         void childrenAlloc(void);
         void insertChild(Node* child, int z);
-        void detachChild(Node *child, ssize_t index, bool doCleanup);
+        void detachChild(Node *child, int64 index, bool doCleanup);
 
         MATH::Vector2f convertToWindowSpace(const MATH::Vector2f& nodePoint) const;
