+++ GRAPH/RENDERER/Renderer.h
@@ -34,13 +34,13 @@ namespace GRAPH
     public:
         RenderQueue();
         void push_back(RenderCommand* command);
-        ssize_t size() const;
+        int64 size() const;
         void sort();
-        RenderCommand* operator[](ssize_t index) const;
+        RenderCommand* operator[](int64 index) const;
         void clear();
         void realloc(size_t reserveSize);
         inline std::vector<RenderCommand*>& getSubQueue(QUEUE_GROUP group) { return commands_[group]; }
-        inline ssize_t getSubQueueSize(QUEUE_GROUP group) const { return commands_[group].size();}
+        inline int64 getSubQueueSize(QUEUE_GROUP group) const { return commands_[group].size();}
 
         void saveRenderState();
         void restoreRenderState();
@@ -56,7 +56,7 @@ namespace GRAPH
     struct RenderStackElement
     {
         int renderQueueID;
-        ssize_t currentIndex;
+        int64 currentIndex;
     };
 
     class GroupCommandManager;
